--- src/Dfe.PlanTech.Domain/Content/Interfaces/ICard.cs
@@ -1,9 +1,0 @@
-﻿namespace Dfe.PlanTech.Domain.Content.Interfaces;
-
-public interface ICard
-{
-    string? Title { get; }
-    string? Description { get; }
-    string? Meta { get; }
-    string? Uri { get; }
-}
