+++ src/Dfe.PlanTech.Domain/Content/Interfaces/ICardContentPartRenderer.cs
@@ -1,8 +1,9 @@
-﻿using Dfe.PlanTech.Domain.Content.Models;
+﻿using System.Text;
+using Dfe.PlanTech.Domain.Content.Models;
 
 namespace Dfe.PlanTech.Domain.Content.Interfaces;
 
-public interface ICardContainerContentPartRenderer
+public interface ICardContentPartRenderer
 {
-    public string ToHtml(List<CsCard>? content);
+    public StringBuilder AddHtml(CsCard? content, StringBuilder stringBuilder);
 }
