+++ src/Dfe.PlanTech.Domain/Content/Interfaces/IRichTextData.cs
@@ -1,6 +1,4 @@
-﻿using Contentful.Core.Models;
-
-namespace Dfe.PlanTech.Domain.Content.Interfaces;
+﻿namespace Dfe.PlanTech.Domain.Content.Interfaces;
 
 /// <summary>
 /// Data for a RichText section
@@ -12,14 +10,4 @@ public interface IRichTextData
     /// URL for a link
     /// </summary>
     public string? Uri { get; init; }
-}
-
-public interface ICardComponentData
-{
-    public string? Title { get; }
-    public string? Description { get; }
-    public string? Meta { get; }
-    public Asset? Image { get; }
-    public string? ImageAlt { get; }
-    public string? Uri { get; }
 }
