+++ src/Dfe.PlanTech.Domain/Content/Models/GridContainer.cs
@@ -5,6 +5,5 @@ namespace Dfe.PlanTech.Domain.Content.Models;
 public class GridContainer : ContentComponent, IContentComponent
 {
     public string? InternalName { get; set; }
-    public List<CsCard> Content { get; set; }
-
+    public IReadOnlyList<CsCard>? Content { get; set; }
 }
