+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/CardComponent.cs
@@ -1,54 +1,29 @@
 ﻿using System.Text;
-using Dfe.PlanTech.Domain.Content.Enums;
 using Dfe.PlanTech.Domain.Content.Interfaces;
 using Dfe.PlanTech.Domain.Content.Models;
 
 namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers;
 
-public class CardComponent : BaseRichTextContentPartRender
+public class CardComponent : ICardContentPartRenderer
 {
-    public CardComponent() : base(RichTextNodeType.CsCard)
+    public CardComponent()
     {
 
     }
 
-    private CustomCard GenerateCustomCard(RichTextContentData content)
+    public StringBuilder AddHtml(CsCard? cardComponent, StringBuilder stringBuilder)
     {
-        return new CustomCard
+        if (cardComponent == null)
         {
-            InternalName = content?.InternalName ?? string.Empty,
-            Description = content?.Asset.Description ?? string.Empty,
-            //check this
-            Meta = content?.Asset.Metadata.ToString(),
-            Uri = content?.Asset.File.Url ?? string.Empty,
-            ImageAlt = string.Empty,
-        };
-    }
-
-    public override StringBuilder AddHtml(RichTextContent content, IRichTextContentPartRendererCollection rendererCollection, StringBuilder stringBuilder)
-    {
-        var target = content?.Data?.Target;
-
-        if (target == null)
-        {
-            return stringBuilder;
+            return null;
         }
 
-        var cardComponent = GenerateCustomCard(target);
-
+        stringBuilder.Append("<div class=\"dfe-card\">");
         stringBuilder.Append("<div class=\"dfe-card-container\">");
         stringBuilder.Append("<h3 class=\"govuk-heading-m\">");
-        stringBuilder.Append($"<a href=\"{cardComponent.Uri}\" class=\"govuk-link govuk-link--no-visited-state dfe-card-link--header\">\"{cardComponent.InternalName}\"</a>");
+        stringBuilder.Append($"<a href=\"{cardComponent.Uri}\" class=\"govuk-link govuk-link--no-visited-state dfe-card-link--header\">{cardComponent.InternalName}</a>");
         stringBuilder.Append("</h3>");
-        if (string.IsNullOrEmpty(cardComponent.Description))
-        {
-            stringBuilder.Append("<p> @Model.Description </p>");
-        }
-        if (string.IsNullOrEmpty(cardComponent.Meta))
-        {
-            stringBuilder.Append($"<p class=\"govuk-body-s>\"{cardComponent.Meta}\"</p>");
-        }
-
+        stringBuilder.Append($"<p class=\"govuk-body-s\">{cardComponent.Description}</p>");
         stringBuilder.Append("</div></div>");
 
         return stringBuilder;
