--- src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/CustomCard.cs
@@ -1,13 +1,0 @@
-﻿namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers
-{
-    internal class CustomCard
-    {
-        public string InternalName { get; set; } = null!;
-        public string Description { get; set; } = null!;
-        public string ImageAlt { get; set; } = null!;
-        public string ImageUri { get; set; } = null!;
-        public string Meta { get; set; } = null!;
-        public string Uri { get; set; } = null!;
-        
-    }
-}
