+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/EmbeddedEntryBlockRenderer.cs
@@ -27,9 +27,6 @@ public class EmbeddedEntryBlockRenderer : BaseRichTextContentPartRender
 
         switch (richTextData.SystemProperties.ContentType?.SystemProperties.Id)
         {
-            case ContentTypeId.Card:
-                var card = new CardComponent();
-                return card.AddHtml(content, rendererCollection, stringBuilder);
             case ContentTypeId.Attachment:
                 var attachment = new AttachmentComponent();
                 return attachment.AddHtml(content, stringBuilder);
