+++ src/Dfe.PlanTech.Infrastructure.Contentful/Helpers/ContentfulSetup.cs
@@ -53,6 +53,7 @@ public static class ContentfulSetup
         services.AddScoped<IRichTextRenderer, RichTextRenderer>();
         services.AddScoped<IRichTextContentPartRendererCollection, RichTextRenderer>();
         services.AddScoped<ICardContainerContentPartRenderer, GridContainerRenderer>();
+        services.AddScoped<ICardContentPartRenderer, CardComponent>();
 
         foreach (var partRenderer in richTextPartRenderers)
         {
